feat: api에서 사용되는 타입을 serviceType으로 이동
