--- client/src/types/createEvent.ts
@@ -1,9 +1,0 @@
-export type EventName = string;
-export type NickName = string;
-export type Password = string;
-
-export type CreateEventArgs = {
-  eventName: EventName;
-  nickname: NickName;
-  password: Password;
-};
